Abstracted some common functions in to common.py
