+++ server/common.py
@@ -1,0 +1,36 @@
+#!/usr/bin/env python3
+
+from sexpdata import Symbol
+import functools
+
+def assoc(key, sexp):
+    if not isinstance(sexp, list):
+        return None
+    for entry in sexp:
+        if isinstance(entry, list) and entry[0] == Symbol(key):
+            return entry[1]
+    return None
+
+def get_body(entry):
+    return entry[-1]
+
+get_user = functools.partial(assoc, "user")
+get_time = functools.partial(assoc, "time")
+get_session = functools.partial(assoc, "session")
+get_id = functools.partial(assoc, "id")
+
+def get_cmd_type(entry):
+    body = get_body(entry)
+    assert body[0] == Symbol("Control")
+    assert isinstance(body[1], list)
+    return body[1][0]
+
+def try_loads(sexp):
+    try:
+        entry = loads(sexp)
+        assert get_user(entry) != None
+        assert get_time(entry)
+        assert get_session(entry)
+        return entry
+    except:
+        return None
