+++ server/replay.py
@@ -8,30 +8,11 @@ import argparse
 from os import listdir
 from os.path import isfile, join
 
+from common import *
+
 logpath = "log.txt"
 logdir = "logs"
 
-def assoc(key, sexp):
-    if not isinstance(sexp, list):
-        return None
-    for entry in sexp:
-        if isinstance(entry, list) and entry[0] == Symbol(key):
-            return entry[1]
-    return None
-
-def get_body(entry):
-    return entry[-1]
-
-get_user = functools.partial(assoc, "user")
-get_time = functools.partial(assoc, "time")
-get_session = functools.partial(assoc, "session")
-get_id = functools.partial(assoc, "id")
-
-def get_cmd_type(entry):
-    body = get_body(entry)
-    assert body[0] == Symbol("Control")
-    assert isinstance(body[1], list)
-    return body[1][0]
 
 class More:
     def __init__(self, num_lines):
@@ -50,16 +31,6 @@ def multipartition(items, f):
             categories[key] = []
         categories[key].append(item)
     return list(categories.values())
-
-def try_loads(sexp):
-    try:
-        entry = loads(sexp)
-        assert get_user(entry) != None
-        assert get_time(entry)
-        assert get_session(entry)
-        return entry
-    except:
-        return None
 
 def main():
     parser = argparse.ArgumentParser()
